Fix Issue 196
If you selected a course during a signin, then unselect it, and tried to
select it again without any other courses selected, The form would be in a locked state where you
couldn't reselect any courses at all.

Formik api call to setFieldValue would remove a form state value if the
value you were trying to set it to was a falsy value. Since when no
courses are selected, we wanted to set it to a empty list, A empty list
is a falsy value resulting in the removal of courses property from form
state causing the form to lock up.

Fixed by bumping formik version and added tests.